+++ src/main/java/firok/tiths/traits/TraitExtremeFreezing.java
@@ -1,0 +1,18 @@
+package firok.tiths.traits;
+
+import slimeknights.tconstruct.library.traits.AbstractTrait;
+
+import static firok.tiths.common.Keys.colorTraitExtremeFreezing;
+import static firok.tiths.common.Keys.nameTraitExtremeFreezing;
+
+// 极寒
+public class TraitExtremeFreezing extends AbstractTrait
+{
+	public TraitExtremeFreezing()
+	{
+		super(nameTraitExtremeFreezing, colorTraitExtremeFreezing);
+	}
+
+	// todo
+	// onUpdate -> freeze mobs
+}